+++ src/ts/dialog.ts
@@ -15,6 +15,14 @@ class Dialog {
     }
 
     element.classList.add("dialog");
+
+    // fecha o diálogo quando o evento de clique é acionado no elemento com o atributo "x-close-dialog".
+    element.querySelector("[x-close-dialog]")
+      .addEventListener("click", e => {
+        this.close();
+
+        e.preventDefault();
+      });
   }
 
   /** A classe CSS de abertura do diálogo. */