+++ src/ts/navigationDrawer.ts
@@ -19,7 +19,16 @@ class NavigationDrawer {
 
     element.classList.add("navigation-drawer");
 
+    // fecha a gaveta de navegação quando o evento de clique é acionado no overlay.
     this._overlay.element.addEventListener("click", () => this.close());
+
+    // fecha a gaveta de navegação quando o evento de clique é acionado no elemento com o atributo "x-close-nav-drawer".
+    element.querySelector("[x-close-nav-drawer]")
+      .addEventListener("click", e => {
+        this.close();
+
+        e.preventDefault();
+      });
   }
 
   /** A classe CSS de abertura da gaveta de navegação. */